Particle swarm optimization

Interface to the particle swarm optimization (pso) algorithm.

Usage

pso(
  env,
  lb,
  ub,
  loss,
  ngen,
  npop,
  error_threshold,
  global = FALSE,
  save_swarm = FALSE,
  run_as_shiny = FALSE,
  add_message = ""
)

Arguments

env

is something that is passed to the loss function in addition to the parameters which get optimized

lb

is a numeric vector defining the lower boundaries of the parameter

ub

is a numeric vector defining the upper boundaries of the parameter

loss

is the loss function for which the optimal parameter set should be found

ngen

is the number of generations the pso should run

npop

is the number of particles which should be used during optimization

error_threshold

is a number defining a sufficient small error at which the optimization is stopped

global

is a logical parameter. If set to TRUE a global star topology is used. Thus, each particle compares itself with the global best particle of the entire swarm. In contrast, if global is set to FALSE a random arbitrary neighborhood is used instead. Thus, each particle has a neighberhood which contains K neighbours where K is between 0 and 3. From the swarm K neighbours are drawn randomly. From the neighberhood the best particle is used for comparison. The neighberhood is calculated for each generation.

save_swarm

is a logical value defining whether the entire optimization should be saved.

run_as_shiny

is an internal parameter which is used when running the shiny app interface.

add_message

is an optional character argument which is printed during optimization

Examples

rosenbrock <- function(parameter, env, Ignore) {
  value <- 0
  for (i in 1:(length(parameter) - 1)) {
    value <- value +
      100 * (parameter[i + 1] - parameter[i]^2)^2 +
      (1 - parameter[i])^2
  }
  return(value)
}

set.seed(1234)
res <- tsf::pso(
  new.env(), rep(-10, 3), rep(10, 3), rosenbrock, 1000, 40,
  0.00001, TRUE, FALSE
)
#> Warning: coercing argument of type 'double' to logical
#> Warning: coercing argument of type 'double' to logical
#> [1] ""
#> [1] 2
#> [1] "-1.165e+00" "-5.885e-01" "1.542e-01" 
#> [1] "3.897e+02"
#> [1] ""
#> [1] 3
#> [1] "-1.881e+00" "1.883e+00"  "4.070e+00" 
#> [1] "3.100e+02"
#> [1] ""
#> [1] 4
#> [1] "-8.802e-01" "1.110e+00"  "2.477e+00" 
#> [1] "1.697e+02"
#> [1] ""
#> [1] 5
#> [1] "-1.441e+00" "1.597e+00"  "2.412e+00" 
#> [1] "3.125e+01"
#> [1] ""
#> [1] 6
#> [1] "-1.441e+00" "1.597e+00"  "2.412e+00" 
#> [1] "3.125e+01"
#> [1] ""
#> [1] 7
#> [1] "-1.441e+00" "1.597e+00"  "2.412e+00" 
#> [1] "3.125e+01"
#> [1] ""
#> [1] 8
#> [1] "-1.441e+00" "1.597e+00"  "2.412e+00" 
#> [1] "3.125e+01"
#> [1] ""
#> [1] 9
#> [1] "-1.441e+00" "1.597e+00"  "2.412e+00" 
#> [1] "3.125e+01"
#> [1] ""
#> [1] 10
#> [1] "-1.441e+00" "1.597e+00"  "2.412e+00" 
#> [1] "3.125e+01"
#> [1] ""
#> [1] 11
#> [1] "-1.441e+00" "1.597e+00"  "2.412e+00" 
#> [1] "3.125e+01"
#> [1] ""
#> [1] 12
#> [1] "-1.441e+00" "1.597e+00"  "2.412e+00" 
#> [1] "3.125e+01"
#> [1] ""
#> [1] 13
#> [1] "-1.441e+00" "1.597e+00"  "2.412e+00" 
#> [1] "3.125e+01"
#> [1] ""
#> [1] 14
#> [1] "-1.441e+00" "1.597e+00"  "2.412e+00" 
#> [1] "3.125e+01"
#> [1] ""
#> [1] 15
#> [1] "-1.441e+00" "1.597e+00"  "2.412e+00" 
#> [1] "3.125e+01"
#> [1] ""
#> [1] 16
#> [1] "-1.170e+00" "1.734e+00"  "2.720e+00" 
#> [1] "2.675e+01"
#> [1] ""
#> [1] 17
#> [1] "-1.170e+00" "1.734e+00"  "2.720e+00" 
#> [1] "2.675e+01"
#> [1] ""
#> [1] 18
#> [1] "-1.170e+00" "1.734e+00"  "2.720e+00" 
#> [1] "2.675e+01"
#> [1] ""
#> [1] 19
#> [1] "-1.170e+00" "1.734e+00"  "2.720e+00" 
#> [1] "2.675e+01"
#> [1] ""
#> [1] 20
#> [1] "-1.170e+00" "1.734e+00"  "2.720e+00" 
#> [1] "2.675e+01"
#> [1] ""
#> [1] 21
#> [1] "-1.170e+00" "1.734e+00"  "2.720e+00" 
#> [1] "2.675e+01"
#> [1] ""
#> [1] 22
#> [1] "-1.170e+00" "1.734e+00"  "2.720e+00" 
#> [1] "2.675e+01"
#> [1] ""
#> [1] 23
#> [1] "-1.170e+00" "1.734e+00"  "2.720e+00" 
#> [1] "2.675e+01"
#> [1] ""
#> [1] 24
#> [1] "-1.170e+00" "1.734e+00"  "2.720e+00" 
#> [1] "2.675e+01"
#> [1] ""
#> [1] 25
#> [1] "-1.170e+00" "1.734e+00"  "2.720e+00" 
#> [1] "2.675e+01"
#> [1] ""
#> [1] 26
#> [1] "-1.170e+00" "1.734e+00"  "2.720e+00" 
#> [1] "2.675e+01"
#> [1] ""
#> [1] 27
#> [1] "-1.170e+00" "1.734e+00"  "2.720e+00" 
#> [1] "2.675e+01"
#> [1] ""
#> [1] 28
#> [1] "-1.170e+00" "1.734e+00"  "2.720e+00" 
#> [1] "2.675e+01"
#> [1] ""
#> [1] 29
#> [1] "-1.170e+00" "1.734e+00"  "2.720e+00" 
#> [1] "2.675e+01"
#> [1] ""
#> [1] 30
#> [1] "-1.170e+00" "1.734e+00"  "2.720e+00" 
#> [1] "2.675e+01"
#> [1] ""
#> [1] 31
#> [1] "-1.170e+00" "1.734e+00"  "2.720e+00" 
#> [1] "2.675e+01"
#> [1] ""
#> [1] 32
#> [1] "-1.170e+00" "1.734e+00"  "2.720e+00" 
#> [1] "2.675e+01"
#> [1] ""
#> [1] 33
#> [1] "-1.170e+00" "1.734e+00"  "2.720e+00" 
#> [1] "2.675e+01"
#> [1] ""
#> [1] 34
#> [1] "-1.170e+00" "1.734e+00"  "2.720e+00" 
#> [1] "2.675e+01"
#> [1] ""
#> [1] 35
#> [1] "-1.170e+00" "1.734e+00"  "2.720e+00" 
#> [1] "2.675e+01"
#> [1] ""
#> [1] 36
#> [1] "-1.170e+00" "1.734e+00"  "2.720e+00" 
#> [1] "2.675e+01"
#> [1] ""
#> [1] 37
#> [1] "-1.170e+00" "1.734e+00"  "2.720e+00" 
#> [1] "2.675e+01"
#> [1] ""
#> [1] 38
#> [1] "-1.170e+00" "1.734e+00"  "2.720e+00" 
#> [1] "2.675e+01"
#> [1] ""
#> [1] 39
#> [1] "-1.170e+00" "1.734e+00"  "2.720e+00" 
#> [1] "2.675e+01"
#> [1] ""
#> [1] 40
#> [1] "-1.170e+00" "1.734e+00"  "2.720e+00" 
#> [1] "2.675e+01"
#> [1] ""
#> [1] 41
#> [1] "-1.170e+00" "1.734e+00"  "2.720e+00" 
#> [1] "2.675e+01"
#> [1] ""
#> [1] 42
#> [1] "-1.340e+00" "1.983e+00"  "4.020e+00" 
#> [1] "1.074e+01"
#> [1] ""
#> [1] 43
#> [1] "-2.782e-02" "-1.514e-01" "1.279e-01" 
#> [1] "5.800e+00"
#> [1] ""
#> [1] 44
#> [1] "-2.782e-02" "-1.514e-01" "1.279e-01" 
#> [1] "5.800e+00"
#> [1] ""
#> [1] 45
#> [1] "-2.782e-02" "-1.514e-01" "1.279e-01" 
#> [1] "5.800e+00"
#> [1] ""
#> [1] 46
#> [1] "-2.782e-02" "-1.514e-01" "1.279e-01" 
#> [1] "5.800e+00"
#> [1] ""
#> [1] 47
#> [1] "-2.782e-02" "-1.514e-01" "1.279e-01" 
#> [1] "5.800e+00"
#> [1] ""
#> [1] 48
#> [1] "-2.782e-02" "-1.514e-01" "1.279e-01" 
#> [1] "5.800e+00"
#> [1] ""
#> [1] 49
#> [1] "-2.782e-02" "-1.514e-01" "1.279e-01" 
#> [1] "5.800e+00"
#> [1] ""
#> [1] 50
#> [1] "-2.782e-02" "-1.514e-01" "1.279e-01" 
#> [1] "5.800e+00"
#> [1] ""
#> [1] 51
#> [1] "-2.782e-02" "-1.514e-01" "1.279e-01" 
#> [1] "5.800e+00"
#> [1] ""
#> [1] 52
#> [1] "-2.782e-02" "-1.514e-01" "1.279e-01" 
#> [1] "5.800e+00"
#> [1] ""
#> [1] 53
#> [1] "-2.782e-02" "-1.514e-01" "1.279e-01" 
#> [1] "5.800e+00"
#> [1] ""
#> [1] 54
#> [1] "-2.782e-02" "-1.514e-01" "1.279e-01" 
#> [1] "5.800e+00"
#> [1] ""
#> [1] 55
#> [1] "-2.782e-02" "-1.514e-01" "1.279e-01" 
#> [1] "5.800e+00"
#> [1] ""
#> [1] 56
#> [1] "-2.782e-02" "-1.514e-01" "1.279e-01" 
#> [1] "5.800e+00"
#> [1] ""
#> [1] 57
#> [1] "-2.782e-02" "-1.514e-01" "1.279e-01" 
#> [1] "5.800e+00"
#> [1] ""
#> [1] 58
#> [1] "-2.782e-02" "-1.514e-01" "1.279e-01" 
#> [1] "5.800e+00"
#> [1] ""
#> [1] 59
#> [1] "-2.782e-02" "-1.514e-01" "1.279e-01" 
#> [1] "5.800e+00"
#> [1] ""
#> [1] 60
#> [1] "-2.782e-02" "-1.514e-01" "1.279e-01" 
#> [1] "5.800e+00"
#> [1] ""
#> [1] 61
#> [1] "-2.782e-02" "-1.514e-01" "1.279e-01" 
#> [1] "5.800e+00"
#> [1] ""
#> [1] 62
#> [1] "-2.782e-02" "-1.514e-01" "1.279e-01" 
#> [1] "5.800e+00"
#> [1] ""
#> [1] 63
#> [1] "-2.782e-02" "-1.514e-01" "1.279e-01" 
#> [1] "5.800e+00"
#> [1] ""
#> [1] 64
#> [1] "-2.782e-02" "-1.514e-01" "1.279e-01" 
#> [1] "5.800e+00"
#> [1] ""
#> [1] 65
#> [1] "-2.782e-02" "-1.514e-01" "1.279e-01" 
#> [1] "5.800e+00"
#> [1] ""
#> [1] 66
#> [1] "-2.782e-02" "-1.514e-01" "1.279e-01" 
#> [1] "5.800e+00"
#> [1] ""
#> [1] 67
#> [1] "-2.782e-02" "-1.514e-01" "1.279e-01" 
#> [1] "5.800e+00"
#> [1] ""
#> [1] 68
#> [1] "-2.782e-02" "-1.514e-01" "1.279e-01" 
#> [1] "5.800e+00"
#> [1] ""
#> [1] 69
#> [1] "-2.782e-02" "-1.514e-01" "1.279e-01" 
#> [1] "5.800e+00"
#> [1] ""
#> [1] 70
#> [1] "-2.782e-02" "-1.514e-01" "1.279e-01" 
#> [1] "5.800e+00"
#> [1] ""
#> [1] 71
#> [1] "-2.782e-02" "-1.514e-01" "1.279e-01" 
#> [1] "5.800e+00"
#> [1] ""
#> [1] 72
#> [1] "-6.759e-01" "3.323e-01"  "3.249e-02" 
#> [1] "5.414e+00"
#> [1] ""
#> [1] 73
#> [1] "-6.759e-01" "3.323e-01"  "3.249e-02" 
#> [1] "5.414e+00"
#> [1] ""
#> [1] 74
#> [1] "-8.014e-01" "7.268e-01"  "4.492e-01" 
#> [1] "4.662e+00"
#> [1] ""
#> [1] 75
#> [1] "4.405e-02"  "-1.001e-01" "1.264e-01" 
#> [1] "4.519e+00"
#> [1] ""
#> [1] 76
#> [1] "4.405e-02"  "-1.001e-01" "1.264e-01" 
#> [1] "4.519e+00"
#> [1] ""
#> [1] 77
#> [1] "-5.016e-01" "2.625e-01"  "1.854e-01" 
#> [1] "4.168e+00"
#> [1] ""
#> [1] 78
#> [1] "-8.205e-01" "6.180e-01"  "3.868e-01" 
#> [1] "3.766e+00"
#> [1] ""
#> [1] 79
#> [1] "-8.205e-01" "6.180e-01"  "3.868e-01" 
#> [1] "3.766e+00"
#> [1] ""
#> [1] 80
#> [1] "-7.331e-01" "5.257e-01"  "3.055e-01" 
#> [1] "3.327e+00"
#> [1] ""
#> [1] 81
#> [1] "-7.331e-01" "5.257e-01"  "3.055e-01" 
#> [1] "3.327e+00"
#> [1] ""
#> [1] 82
#> [1] "-7.331e-01" "5.257e-01"  "3.055e-01" 
#> [1] "3.327e+00"
#> [1] ""
#> [1] 83
#> [1] "-7.331e-01" "5.257e-01"  "3.055e-01" 
#> [1] "3.327e+00"
#> [1] ""
#> [1] 84
#> [1] "5.716e-01" "2.335e-01" "8.583e-02"
#> [1] "1.740e+00"
#> [1] ""
#> [1] 85
#> [1] "5.716e-01" "2.335e-01" "8.583e-02"
#> [1] "1.740e+00"
#> [1] ""
#> [1] 86
#> [1] "6.937e-01" "4.349e-01" "1.326e-01"
#> [1] "9.473e-01"
#> [1] ""
#> [1] 87
#> [1] "6.937e-01" "4.349e-01" "1.326e-01"
#> [1] "9.473e-01"
#> [1] ""
#> [1] 88
#> [1] "6.551e-01" "4.296e-01" "1.212e-01"
#> [1] "8.463e-01"
#> [1] ""
#> [1] 89
#> [1] "6.551e-01" "4.296e-01" "1.212e-01"
#> [1] "8.463e-01"
#> [1] ""
#> [1] 90
#> [1] "6.551e-01" "4.296e-01" "1.212e-01"
#> [1] "8.463e-01"
#> [1] ""
#> [1] 91
#> [1] "6.551e-01" "4.296e-01" "1.212e-01"
#> [1] "8.463e-01"
#> [1] ""
#> [1] 92
#> [1] "6.551e-01" "4.296e-01" "1.212e-01"
#> [1] "8.463e-01"
#> [1] ""
#> [1] 93
#> [1] "6.551e-01" "4.296e-01" "1.212e-01"
#> [1] "8.463e-01"
#> [1] ""
#> [1] 94
#> [1] "6.551e-01" "4.296e-01" "1.212e-01"
#> [1] "8.463e-01"
#> [1] ""
#> [1] 95
#> [1] "6.551e-01" "4.296e-01" "1.212e-01"
#> [1] "8.463e-01"
#> [1] ""
#> [1] 96
#> [1] "6.551e-01" "4.296e-01" "1.212e-01"
#> [1] "8.463e-01"
#> [1] ""
#> [1] 97
#> [1] "6.551e-01" "4.296e-01" "1.212e-01"
#> [1] "8.463e-01"
#> [1] ""
#> [1] 98
#> [1] "6.551e-01" "4.296e-01" "1.212e-01"
#> [1] "8.463e-01"
#> [1] ""
#> [1] 99
#> [1] "6.551e-01" "4.296e-01" "1.212e-01"
#> [1] "8.463e-01"
#> [1] ""
#> [1] 100
#> [1] "6.551e-01" "4.296e-01" "1.212e-01"
#> [1] "8.463e-01"
#> [1] ""
#> [1] 101
#> [1] "6.551e-01" "4.296e-01" "1.212e-01"
#> [1] "8.463e-01"
#> [1] ""
#> [1] 102
#> [1] "6.551e-01" "4.296e-01" "1.212e-01"
#> [1] "8.463e-01"
#> [1] ""
#> [1] 103
#> [1] "6.551e-01" "4.296e-01" "1.212e-01"
#> [1] "8.463e-01"
#> [1] ""
#> [1] 104
#> [1] "6.551e-01" "4.296e-01" "1.212e-01"
#> [1] "8.463e-01"
#> [1] ""
#> [1] 105
#> [1] "6.551e-01" "4.296e-01" "1.212e-01"
#> [1] "8.463e-01"
#> [1] ""
#> [1] 106
#> [1] "6.551e-01" "4.296e-01" "1.212e-01"
#> [1] "8.463e-01"
#> [1] ""
#> [1] 107
#> [1] "6.551e-01" "4.296e-01" "1.212e-01"
#> [1] "8.463e-01"
#> [1] ""
#> [1] 108
#> [1] "6.551e-01" "4.296e-01" "1.212e-01"
#> [1] "8.463e-01"
#> [1] ""
#> [1] 109
#> [1] "6.551e-01" "4.296e-01" "1.212e-01"
#> [1] "8.463e-01"
#> [1] ""
#> [1] 110
#> [1] "6.551e-01" "4.296e-01" "1.212e-01"
#> [1] "8.463e-01"
#> [1] ""
#> [1] 111
#> [1] "6.551e-01" "4.296e-01" "1.212e-01"
#> [1] "8.463e-01"
#> [1] ""
#> [1] 112
#> [1] "6.551e-01" "4.296e-01" "1.212e-01"
#> [1] "8.463e-01"
#> [1] ""
#> [1] 113
#> [1] "6.551e-01" "4.296e-01" "1.212e-01"
#> [1] "8.463e-01"
#> [1] ""
#> [1] 114
#> [1] "6.551e-01" "4.296e-01" "1.212e-01"
#> [1] "8.463e-01"
#> [1] ""
#> [1] 115
#> [1] "6.551e-01" "4.296e-01" "1.212e-01"
#> [1] "8.463e-01"
#> [1] ""
#> [1] 116
#> [1] "6.551e-01" "4.296e-01" "1.212e-01"
#> [1] "8.463e-01"
#> [1] ""
#> [1] 117
#> [1] "6.551e-01" "4.296e-01" "1.212e-01"
#> [1] "8.463e-01"
#> [1] ""
#> [1] 118
#> [1] "6.551e-01" "4.296e-01" "1.212e-01"
#> [1] "8.463e-01"
#> [1] ""
#> [1] 119
#> [1] "6.551e-01" "4.296e-01" "1.212e-01"
#> [1] "8.463e-01"
#> [1] ""
#> [1] 120
#> [1] "6.551e-01" "4.296e-01" "1.212e-01"
#> [1] "8.463e-01"
#> [1] ""
#> [1] 121
#> [1] "7.474e-01" "5.475e-01" "2.572e-01"
#> [1] "4.624e-01"
#> [1] ""
#> [1] 122
#> [1] "7.474e-01" "5.475e-01" "2.572e-01"
#> [1] "4.624e-01"
#> [1] ""
#> [1] 123
#> [1] "7.474e-01" "5.475e-01" "2.572e-01"
#> [1] "4.624e-01"
#> [1] ""
#> [1] 124
#> [1] "7.474e-01" "5.475e-01" "2.572e-01"
#> [1] "4.624e-01"
#> [1] ""
#> [1] 125
#> [1] "7.474e-01" "5.475e-01" "2.572e-01"
#> [1] "4.624e-01"
#> [1] ""
#> [1] 126
#> [1] "7.474e-01" "5.475e-01" "2.572e-01"
#> [1] "4.624e-01"
#> [1] ""
#> [1] 127
#> [1] "7.474e-01" "5.475e-01" "2.572e-01"
#> [1] "4.624e-01"
#> [1] ""
#> [1] 128
#> [1] "7.474e-01" "5.475e-01" "2.572e-01"
#> [1] "4.624e-01"
#> [1] ""
#> [1] 129
#> [1] "7.474e-01" "5.475e-01" "2.572e-01"
#> [1] "4.624e-01"
#> [1] ""
#> [1] 130
#> [1] "7.474e-01" "5.475e-01" "2.572e-01"
#> [1] "4.624e-01"
#> [1] ""
#> [1] 131
#> [1] "7.474e-01" "5.475e-01" "2.572e-01"
#> [1] "4.624e-01"
#> [1] ""
#> [1] 132
#> [1] "7.474e-01" "5.475e-01" "2.572e-01"
#> [1] "4.624e-01"
#> [1] ""
#> [1] 133
#> [1] "7.474e-01" "5.475e-01" "2.572e-01"
#> [1] "4.624e-01"
#> [1] ""
#> [1] 134
#> [1] "7.474e-01" "5.475e-01" "2.572e-01"
#> [1] "4.624e-01"
#> [1] ""
#> [1] 135
#> [1] "7.474e-01" "5.475e-01" "2.572e-01"
#> [1] "4.624e-01"
#> [1] ""
#> [1] 136
#> [1] "7.474e-01" "5.475e-01" "2.572e-01"
#> [1] "4.624e-01"
#> [1] ""
#> [1] 137
#> [1] "7.474e-01" "5.475e-01" "2.572e-01"
#> [1] "4.624e-01"
#> [1] ""
#> [1] 138
#> [1] "7.474e-01" "5.475e-01" "2.572e-01"
#> [1] "4.624e-01"
#> [1] ""
#> [1] 139
#> [1] "7.474e-01" "5.475e-01" "2.572e-01"
#> [1] "4.624e-01"
#> [1] ""
#> [1] 140
#> [1] "7.474e-01" "5.475e-01" "2.572e-01"
#> [1] "4.624e-01"
#> [1] ""
#> [1] 141
#> [1] "7.474e-01" "5.475e-01" "2.572e-01"
#> [1] "4.624e-01"
#> [1] ""
#> [1] 142
#> [1] "7.474e-01" "5.475e-01" "2.572e-01"
#> [1] "4.624e-01"
#> [1] ""
#> [1] 143
#> [1] "7.474e-01" "5.475e-01" "2.572e-01"
#> [1] "4.624e-01"
#> [1] ""
#> [1] 144
#> [1] "7.474e-01" "5.475e-01" "2.572e-01"
#> [1] "4.624e-01"
#> [1] ""
#> [1] 145
#> [1] "7.474e-01" "5.475e-01" "2.572e-01"
#> [1] "4.624e-01"
#> [1] ""
#> [1] 146
#> [1] "7.474e-01" "5.475e-01" "2.572e-01"
#> [1] "4.624e-01"
#> [1] ""
#> [1] 147
#> [1] "7.474e-01" "5.475e-01" "2.572e-01"
#> [1] "4.624e-01"
#> [1] ""
#> [1] 148
#> [1] "7.474e-01" "5.475e-01" "2.572e-01"
#> [1] "4.624e-01"
#> [1] ""
#> [1] 149
#> [1] "7.474e-01" "5.475e-01" "2.572e-01"
#> [1] "4.624e-01"
#> [1] ""
#> [1] 150
#> [1] "7.474e-01" "5.475e-01" "2.572e-01"
#> [1] "4.624e-01"
#> [1] ""
#> [1] 151
#> [1] "7.474e-01" "5.475e-01" "2.572e-01"
#> [1] "4.624e-01"
#> [1] ""
#> [1] 152
#> [1] "7.474e-01" "5.475e-01" "2.572e-01"
#> [1] "4.624e-01"
#> [1] ""
#> [1] 153
#> [1] "7.474e-01" "5.475e-01" "2.572e-01"
#> [1] "4.624e-01"
#> [1] ""
#> [1] 154
#> [1] "7.474e-01" "5.475e-01" "2.572e-01"
#> [1] "4.624e-01"
#> [1] ""
#> [1] 155
#> [1] "7.474e-01" "5.475e-01" "2.572e-01"
#> [1] "4.624e-01"
#> [1] ""
#> [1] 156
#> [1] "7.474e-01" "5.475e-01" "2.572e-01"
#> [1] "4.624e-01"
#> [1] ""
#> [1] 157
#> [1] "7.474e-01" "5.475e-01" "2.572e-01"
#> [1] "4.624e-01"
#> [1] ""
#> [1] 158
#> [1] "7.474e-01" "5.475e-01" "2.572e-01"
#> [1] "4.624e-01"
#> [1] ""
#> [1] 159
#> [1] "7.851e-01" "5.765e-01" "3.406e-01"
#> [1] "3.907e-01"
#> [1] ""
#> [1] 160
#> [1] "7.851e-01" "5.765e-01" "3.406e-01"
#> [1] "3.907e-01"
#> [1] ""
#> [1] 161
#> [1] "7.851e-01" "5.765e-01" "3.406e-01"
#> [1] "3.907e-01"
#> [1] ""
#> [1] 162
#> [1] "7.851e-01" "5.765e-01" "3.406e-01"
#> [1] "3.907e-01"
#> [1] ""
#> [1] 163
#> [1] "7.851e-01" "5.765e-01" "3.406e-01"
#> [1] "3.907e-01"
#> [1] ""
#> [1] 164
#> [1] "7.851e-01" "5.765e-01" "3.406e-01"
#> [1] "3.907e-01"
#> [1] ""
#> [1] 165
#> [1] "7.851e-01" "5.765e-01" "3.406e-01"
#> [1] "3.907e-01"
#> [1] ""
#> [1] 166
#> [1] "7.851e-01" "5.765e-01" "3.406e-01"
#> [1] "3.907e-01"
#> [1] ""
#> [1] 167
#> [1] "7.851e-01" "5.765e-01" "3.406e-01"
#> [1] "3.907e-01"
#> [1] ""
#> [1] 168
#> [1] "7.851e-01" "5.765e-01" "3.406e-01"
#> [1] "3.907e-01"
#> [1] ""
#> [1] 169
#> [1] "7.851e-01" "5.765e-01" "3.406e-01"
#> [1] "3.907e-01"
#> [1] ""
#> [1] 170
#> [1] "7.851e-01" "5.765e-01" "3.406e-01"
#> [1] "3.907e-01"
#> [1] ""
#> [1] 171
#> [1] "7.851e-01" "5.765e-01" "3.406e-01"
#> [1] "3.907e-01"
#> [1] ""
#> [1] 172
#> [1] "7.851e-01" "5.765e-01" "3.406e-01"
#> [1] "3.907e-01"
#> [1] ""
#> [1] 173
#> [1] "7.851e-01" "5.765e-01" "3.406e-01"
#> [1] "3.907e-01"
#> [1] ""
#> [1] 174
#> [1] "7.851e-01" "5.765e-01" "3.406e-01"
#> [1] "3.907e-01"
#> [1] ""
#> [1] 175
#> [1] "7.851e-01" "5.765e-01" "3.406e-01"
#> [1] "3.907e-01"
#> [1] ""
#> [1] 176
#> [1] "7.851e-01" "5.765e-01" "3.406e-01"
#> [1] "3.907e-01"
#> [1] ""
#> [1] 177
#> [1] "7.851e-01" "5.765e-01" "3.406e-01"
#> [1] "3.907e-01"
#> [1] ""
#> [1] 178
#> [1] "7.851e-01" "5.765e-01" "3.406e-01"
#> [1] "3.907e-01"
#> [1] ""
#> [1] 179
#> [1] "7.851e-01" "5.765e-01" "3.406e-01"
#> [1] "3.907e-01"
#> [1] ""
#> [1] 180
#> [1] "7.851e-01" "5.765e-01" "3.406e-01"
#> [1] "3.907e-01"
#> [1] ""
#> [1] 181
#> [1] "7.851e-01" "5.765e-01" "3.406e-01"
#> [1] "3.907e-01"
#> [1] ""
#> [1] 182
#> [1] "7.851e-01" "5.765e-01" "3.406e-01"
#> [1] "3.907e-01"
#> [1] ""
#> [1] 183
#> [1] "7.851e-01" "5.765e-01" "3.406e-01"
#> [1] "3.907e-01"
#> [1] ""
#> [1] 184
#> [1] "7.851e-01" "5.765e-01" "3.406e-01"
#> [1] "3.907e-01"
#> [1] ""
#> [1] 185
#> [1] "7.851e-01" "5.765e-01" "3.406e-01"
#> [1] "3.907e-01"
#> [1] ""
#> [1] 186
#> [1] "7.851e-01" "5.765e-01" "3.406e-01"
#> [1] "3.907e-01"
#> [1] ""
#> [1] 187
#> [1] "7.851e-01" "5.765e-01" "3.406e-01"
#> [1] "3.907e-01"
#> [1] ""
#> [1] 188
#> [1] "7.851e-01" "5.765e-01" "3.406e-01"
#> [1] "3.907e-01"
#> [1] ""
#> [1] 189
#> [1] "7.851e-01" "5.765e-01" "3.406e-01"
#> [1] "3.907e-01"
#> [1] ""
#> [1] 190
#> [1] "7.851e-01" "5.765e-01" "3.406e-01"
#> [1] "3.907e-01"
#> [1] ""
#> [1] 191
#> [1] "7.851e-01" "5.765e-01" "3.406e-01"
#> [1] "3.907e-01"
#> [1] ""
#> [1] 192
#> [1] "7.851e-01" "5.765e-01" "3.406e-01"
#> [1] "3.907e-01"
#> [1] ""
#> [1] 193
#> [1] "7.851e-01" "5.765e-01" "3.406e-01"
#> [1] "3.907e-01"
#> [1] ""
#> [1] 194
#> [1] "7.851e-01" "5.765e-01" "3.406e-01"
#> [1] "3.907e-01"
#> [1] ""
#> [1] 195
#> [1] "7.851e-01" "5.765e-01" "3.406e-01"
#> [1] "3.907e-01"
#> [1] ""
#> [1] 196
#> [1] "7.851e-01" "5.765e-01" "3.406e-01"
#> [1] "3.907e-01"
#> [1] ""
#> [1] 197
#> [1] "7.851e-01" "5.765e-01" "3.406e-01"
#> [1] "3.907e-01"
#> [1] ""
#> [1] 198
#> [1] "7.851e-01" "5.765e-01" "3.406e-01"
#> [1] "3.907e-01"
#> [1] ""
#> [1] 199
#> [1] "7.851e-01" "5.765e-01" "3.406e-01"
#> [1] "3.907e-01"
#> [1] ""
#> [1] 200
#> [1] "7.851e-01" "5.765e-01" "3.406e-01"
#> [1] "3.907e-01"
#> [1] ""
#> [1] 201
#> [1] "7.851e-01" "5.765e-01" "3.406e-01"
#> [1] "3.907e-01"
#> [1] ""
#> [1] 202
#> [1] "7.851e-01" "5.765e-01" "3.406e-01"
#> [1] "3.907e-01"
#> [1] ""
#> [1] 203
#> [1] "7.851e-01" "5.765e-01" "3.406e-01"
#> [1] "3.907e-01"
#> [1] ""
#> [1] 204
#> [1] "7.851e-01" "5.765e-01" "3.406e-01"
#> [1] "3.907e-01"
#> [1] ""
#> [1] 205
#> [1] "7.851e-01" "5.765e-01" "3.406e-01"
#> [1] "3.907e-01"
#> [1] ""
#> [1] 206
#> [1] "7.851e-01" "5.765e-01" "3.406e-01"
#> [1] "3.907e-01"
#> [1] ""
#> [1] 207
#> [1] "7.851e-01" "5.765e-01" "3.406e-01"
#> [1] "3.907e-01"
#> [1] ""
#> [1] 208
#> [1] "7.851e-01" "5.765e-01" "3.406e-01"
#> [1] "3.907e-01"
#> [1] ""
#> [1] 209
#> [1] "7.851e-01" "5.765e-01" "3.406e-01"
#> [1] "3.907e-01"
#> [1] ""
#> [1] 210
#> [1] "9.013e-01" "7.919e-01" "6.624e-01"
#> [1] "2.193e-01"
#> [1] ""
#> [1] 211
#> [1] "9.013e-01" "7.919e-01" "6.624e-01"
#> [1] "2.193e-01"
#> [1] ""
#> [1] 212
#> [1] "9.013e-01" "7.919e-01" "6.624e-01"
#> [1] "2.193e-01"
#> [1] ""
#> [1] 213
#> [1] "9.013e-01" "7.919e-01" "6.624e-01"
#> [1] "2.193e-01"
#> [1] ""
#> [1] 214
#> [1] "9.013e-01" "7.919e-01" "6.624e-01"
#> [1] "2.193e-01"
#> [1] ""
#> [1] 215
#> [1] "9.013e-01" "7.919e-01" "6.624e-01"
#> [1] "2.193e-01"
#> [1] ""
#> [1] 216
#> [1] "9.013e-01" "7.919e-01" "6.624e-01"
#> [1] "2.193e-01"
#> [1] ""
#> [1] 217
#> [1] "9.013e-01" "7.919e-01" "6.624e-01"
#> [1] "2.193e-01"
#> [1] ""
#> [1] 218
#> [1] "9.013e-01" "7.919e-01" "6.624e-01"
#> [1] "2.193e-01"
#> [1] ""
#> [1] 219
#> [1] "9.013e-01" "7.919e-01" "6.624e-01"
#> [1] "2.193e-01"
#> [1] ""
#> [1] 220
#> [1] "9.013e-01" "7.919e-01" "6.624e-01"
#> [1] "2.193e-01"
#> [1] ""
#> [1] 221
#> [1] "9.013e-01" "7.919e-01" "6.624e-01"
#> [1] "2.193e-01"
#> [1] ""
#> [1] 222
#> [1] "9.013e-01" "7.919e-01" "6.624e-01"
#> [1] "2.193e-01"
#> [1] ""
#> [1] 223
#> [1] "9.013e-01" "7.919e-01" "6.624e-01"
#> [1] "2.193e-01"
#> [1] ""
#> [1] 224
#> [1] "9.013e-01" "7.919e-01" "6.624e-01"
#> [1] "2.193e-01"
#> [1] ""
#> [1] 225
#> [1] "9.013e-01" "7.919e-01" "6.624e-01"
#> [1] "2.193e-01"
#> [1] ""
#> [1] 226
#> [1] "9.013e-01" "7.919e-01" "6.624e-01"
#> [1] "2.193e-01"
#> [1] ""
#> [1] 227
#> [1] "9.013e-01" "7.919e-01" "6.624e-01"
#> [1] "2.193e-01"
#> [1] ""
#> [1] 228
#> [1] "9.036e-01" "7.942e-01" "6.646e-01"
#> [1] "2.159e-01"
#> [1] ""
#> [1] 229
#> [1] "9.036e-01" "7.942e-01" "6.646e-01"
#> [1] "2.159e-01"
#> [1] ""
#> [1] 230
#> [1] "9.036e-01" "7.942e-01" "6.646e-01"
#> [1] "2.159e-01"
#> [1] ""
#> [1] 231
#> [1] "9.036e-01" "7.942e-01" "6.646e-01"
#> [1] "2.159e-01"
#> [1] ""
#> [1] 232
#> [1] "9.036e-01" "7.942e-01" "6.646e-01"
#> [1] "2.159e-01"
#> [1] ""
#> [1] 233
#> [1] "9.036e-01" "7.942e-01" "6.646e-01"
#> [1] "2.159e-01"
#> [1] ""
#> [1] 234
#> [1] "9.036e-01" "7.942e-01" "6.646e-01"
#> [1] "2.159e-01"
#> [1] ""
#> [1] 235
#> [1] "9.036e-01" "7.942e-01" "6.646e-01"
#> [1] "2.159e-01"
#> [1] ""
#> [1] 236
#> [1] "9.036e-01" "7.942e-01" "6.646e-01"
#> [1] "2.159e-01"
#> [1] ""
#> [1] 237
#> [1] "9.036e-01" "7.942e-01" "6.646e-01"
#> [1] "2.159e-01"
#> [1] ""
#> [1] 238
#> [1] "9.036e-01" "7.942e-01" "6.646e-01"
#> [1] "2.159e-01"
#> [1] ""
#> [1] 239
#> [1] "9.036e-01" "7.942e-01" "6.646e-01"
#> [1] "2.159e-01"
#> [1] ""
#> [1] 240
#> [1] "9.036e-01" "7.942e-01" "6.646e-01"
#> [1] "2.159e-01"
#> [1] ""
#> [1] 241
#> [1] "9.036e-01" "7.942e-01" "6.646e-01"
#> [1] "2.159e-01"
#> [1] ""
#> [1] 242
#> [1] "9.036e-01" "7.942e-01" "6.646e-01"
#> [1] "2.159e-01"
#> [1] ""
#> [1] 243
#> [1] "9.036e-01" "7.942e-01" "6.646e-01"
#> [1] "2.159e-01"
#> [1] ""
#> [1] 244
#> [1] "9.036e-01" "7.942e-01" "6.646e-01"
#> [1] "2.159e-01"
#> [1] ""
#> [1] 245
#> [1] "9.036e-01" "7.942e-01" "6.646e-01"
#> [1] "2.159e-01"
#> [1] ""
#> [1] 246
#> [1] "9.036e-01" "7.942e-01" "6.646e-01"
#> [1] "2.159e-01"
#> [1] ""
#> [1] 247
#> [1] "9.036e-01" "7.942e-01" "6.646e-01"
#> [1] "2.159e-01"
#> [1] ""
#> [1] 248
#> [1] "1.011e+00" "9.857e-01" "9.473e-01"
#> [1] "1.878e-01"
#> [1] ""
#> [1] 249
#> [1] "1.011e+00" "9.857e-01" "9.473e-01"
#> [1] "1.878e-01"
#> [1] ""
#> [1] 250
#> [1] "1.014e+00" "9.993e-01" "9.688e-01"
#> [1] "1.769e-01"
#> [1] ""
#> [1] 251
#> [1] "1.014e+00" "9.993e-01" "9.688e-01"
#> [1] "1.769e-01"
#> [1] ""
#> [1] 252
#> [1] "1.014e+00" "9.993e-01" "9.688e-01"
#> [1] "1.769e-01"
#> [1] ""
#> [1] 253
#> [1] "1.014e+00" "9.993e-01" "9.688e-01"
#> [1] "1.769e-01"
#> [1] ""
#> [1] 254
#> [1] "1.014e+00" "9.993e-01" "9.688e-01"
#> [1] "1.769e-01"
#> [1] ""
#> [1] 255
#> [1] "1.014e+00" "9.993e-01" "9.688e-01"
#> [1] "1.769e-01"
#> [1] ""
#> [1] 256
#> [1] "1.014e+00" "9.993e-01" "9.688e-01"
#> [1] "1.769e-01"
#> [1] ""
#> [1] 257
#> [1] "1.010e+00" "1.007e+00" "9.896e-01"
#> [1] "7.686e-02"
#> [1] ""
#> [1] 258
#> [1] "1.010e+00" "1.007e+00" "9.896e-01"
#> [1] "7.686e-02"
#> [1] ""
#> [1] 259
#> [1] "1.010e+00" "1.007e+00" "9.896e-01"
#> [1] "7.686e-02"
#> [1] ""
#> [1] 260
#> [1] "1.010e+00" "1.007e+00" "9.896e-01"
#> [1] "7.686e-02"
#> [1] ""
#> [1] 261
#> [1] "1.010e+00" "1.007e+00" "9.896e-01"
#> [1] "7.686e-02"
#> [1] ""
#> [1] 262
#> [1] "1.010e+00" "1.007e+00" "9.896e-01"
#> [1] "7.686e-02"
#> [1] ""
#> [1] 263
#> [1] "1.010e+00" "1.007e+00" "9.896e-01"
#> [1] "7.686e-02"
#> [1] ""
#> [1] 264
#> [1] "1.010e+00" "1.007e+00" "9.896e-01"
#> [1] "7.686e-02"
#> [1] ""
#> [1] 265
#> [1] "9.848e-01" "9.770e-01" "9.581e-01"
#> [1] "7.234e-03"
#> [1] ""
#> [1] 266
#> [1] "9.848e-01" "9.770e-01" "9.581e-01"
#> [1] "7.234e-03"
#> [1] ""
#> [1] 267
#> [1] "9.873e-01" "9.797e-01" "9.606e-01"
#> [1] "3.126e-03"
#> [1] ""
#> [1] 268
#> [1] "9.873e-01" "9.797e-01" "9.606e-01"
#> [1] "3.126e-03"
#> [1] ""
#> [1] 269
#> [1] "9.873e-01" "9.797e-01" "9.606e-01"
#> [1] "3.126e-03"
#> [1] ""
#> [1] 270
#> [1] "9.873e-01" "9.797e-01" "9.606e-01"
#> [1] "3.126e-03"
#> [1] ""
#> [1] 271
#> [1] "9.873e-01" "9.797e-01" "9.606e-01"
#> [1] "3.126e-03"
#> [1] ""
#> [1] 272
#> [1] "9.873e-01" "9.797e-01" "9.606e-01"
#> [1] "3.126e-03"
#> [1] ""
#> [1] 273
#> [1] "9.873e-01" "9.797e-01" "9.606e-01"
#> [1] "3.126e-03"
#> [1] ""
#> [1] 274
#> [1] "9.873e-01" "9.797e-01" "9.606e-01"
#> [1] "3.126e-03"
#> [1] ""
#> [1] 275
#> [1] "9.873e-01" "9.797e-01" "9.606e-01"
#> [1] "3.126e-03"
#> [1] ""
#> [1] 276
#> [1] "9.873e-01" "9.797e-01" "9.606e-01"
#> [1] "3.126e-03"
#> [1] ""
#> [1] 277
#> [1] "9.873e-01" "9.797e-01" "9.606e-01"
#> [1] "3.126e-03"
#> [1] ""
#> [1] 278
#> [1] "9.873e-01" "9.797e-01" "9.606e-01"
#> [1] "3.126e-03"
#> [1] ""
#> [1] 279
#> [1] "9.873e-01" "9.797e-01" "9.606e-01"
#> [1] "3.126e-03"
#> [1] ""
#> [1] 280
#> [1] "9.873e-01" "9.797e-01" "9.606e-01"
#> [1] "3.126e-03"
#> [1] ""
#> [1] 281
#> [1] "9.873e-01" "9.797e-01" "9.606e-01"
#> [1] "3.126e-03"
#> [1] ""
#> [1] 282
#> [1] "9.873e-01" "9.797e-01" "9.606e-01"
#> [1] "3.126e-03"
#> [1] ""
#> [1] 283
#> [1] "9.873e-01" "9.797e-01" "9.606e-01"
#> [1] "3.126e-03"
#> [1] ""
#> [1] 284
#> [1] "9.873e-01" "9.797e-01" "9.606e-01"
#> [1] "3.126e-03"
#> [1] ""
#> [1] 285
#> [1] "9.873e-01" "9.797e-01" "9.606e-01"
#> [1] "3.126e-03"
#> [1] ""
#> [1] 286
#> [1] "9.873e-01" "9.797e-01" "9.606e-01"
#> [1] "3.126e-03"
#> [1] ""
#> [1] 287
#> [1] "9.873e-01" "9.797e-01" "9.606e-01"
#> [1] "3.126e-03"
#> [1] ""
#> [1] 288
#> [1] "9.873e-01" "9.797e-01" "9.606e-01"
#> [1] "3.126e-03"
#> [1] ""
#> [1] 289
#> [1] "9.873e-01" "9.797e-01" "9.606e-01"
#> [1] "3.126e-03"
#> [1] ""
#> [1] 290
#> [1] "9.873e-01" "9.797e-01" "9.606e-01"
#> [1] "3.126e-03"
#> [1] ""
#> [1] 291
#> [1] "9.873e-01" "9.797e-01" "9.606e-01"
#> [1] "3.126e-03"
#> [1] ""
#> [1] 292
#> [1] "9.873e-01" "9.797e-01" "9.606e-01"
#> [1] "3.126e-03"
#> [1] ""
#> [1] 293
#> [1] "9.873e-01" "9.797e-01" "9.606e-01"
#> [1] "3.126e-03"
#> [1] ""
#> [1] 294
#> [1] "9.873e-01" "9.797e-01" "9.606e-01"
#> [1] "3.126e-03"
#> [1] ""
#> [1] 295
#> [1] "9.873e-01" "9.797e-01" "9.606e-01"
#> [1] "3.126e-03"
#> [1] ""
#> [1] 296
#> [1] "9.873e-01" "9.797e-01" "9.606e-01"
#> [1] "3.126e-03"
#> [1] ""
#> [1] 297
#> [1] "9.873e-01" "9.797e-01" "9.606e-01"
#> [1] "3.126e-03"
#> [1] ""
#> [1] 298
#> [1] "9.873e-01" "9.797e-01" "9.606e-01"
#> [1] "3.126e-03"
#> [1] ""
#> [1] 299
#> [1] "9.873e-01" "9.797e-01" "9.606e-01"
#> [1] "3.126e-03"
#> [1] ""
#> [1] 300
#> [1] "9.911e-01" "9.848e-01" "9.665e-01"
#> [1] "2.038e-03"
#> [1] ""
#> [1] 301
#> [1] "9.911e-01" "9.848e-01" "9.665e-01"
#> [1] "2.038e-03"
#> [1] ""
#> [1] 302
#> [1] "9.911e-01" "9.848e-01" "9.665e-01"
#> [1] "2.038e-03"
#> [1] ""
#> [1] 303
#> [1] "9.911e-01" "9.848e-01" "9.665e-01"
#> [1] "2.038e-03"
#> [1] ""
#> [1] 304
#> [1] "9.911e-01" "9.848e-01" "9.665e-01"
#> [1] "2.038e-03"
#> [1] ""
#> [1] 305
#> [1] "9.911e-01" "9.848e-01" "9.665e-01"
#> [1] "2.038e-03"
#> [1] ""
#> [1] 306
#> [1] "9.911e-01" "9.848e-01" "9.665e-01"
#> [1] "2.038e-03"
#> [1] ""
#> [1] 307
#> [1] "9.911e-01" "9.848e-01" "9.665e-01"
#> [1] "2.038e-03"
#> [1] ""
#> [1] 308
#> [1] "9.911e-01" "9.848e-01" "9.665e-01"
#> [1] "2.038e-03"
#> [1] ""
#> [1] 309
#> [1] "9.911e-01" "9.848e-01" "9.665e-01"
#> [1] "2.038e-03"
#> [1] ""
#> [1] 310
#> [1] "9.910e-01" "9.847e-01" "9.664e-01"
#> [1] "2.012e-03"
#> [1] ""
#> [1] 311
#> [1] "9.911e-01" "9.814e-01" "9.624e-01"
#> [1] "5.499e-04"
#> [1] ""
#> [1] 312
#> [1] "9.904e-01" "9.810e-01" "9.627e-01"
#> [1] "4.704e-04"
#> [1] ""
#> [1] 313
#> [1] "9.904e-01" "9.810e-01" "9.627e-01"
#> [1] "4.704e-04"
#> [1] ""
#> [1] 314
#> [1] "9.904e-01" "9.810e-01" "9.627e-01"
#> [1] "4.704e-04"
#> [1] ""
#> [1] 315
#> [1] "9.915e-01" "9.834e-01" "9.666e-01"
#> [1] "3.868e-04"
#> [1] ""
#> [1] 316
#> [1] "9.915e-01" "9.834e-01" "9.666e-01"
#> [1] "3.868e-04"
#> [1] ""
#> [1] 317
#> [1] "9.915e-01" "9.834e-01" "9.666e-01"
#> [1] "3.868e-04"
#> [1] ""
#> [1] 318
#> [1] "9.915e-01" "9.834e-01" "9.666e-01"
#> [1] "3.868e-04"
#> [1] ""
#> [1] 319
#> [1] "9.915e-01" "9.834e-01" "9.666e-01"
#> [1] "3.868e-04"
#> [1] ""
#> [1] 320
#> [1] "9.915e-01" "9.834e-01" "9.666e-01"
#> [1] "3.868e-04"
#> [1] ""
#> [1] 321
#> [1] "9.915e-01" "9.834e-01" "9.666e-01"
#> [1] "3.868e-04"
#> [1] ""
#> [1] 322
#> [1] "9.915e-01" "9.834e-01" "9.666e-01"
#> [1] "3.868e-04"
#> [1] ""
#> [1] 323
#> [1] "9.915e-01" "9.834e-01" "9.666e-01"
#> [1] "3.868e-04"
#> [1] ""
#> [1] 324
#> [1] "9.915e-01" "9.834e-01" "9.666e-01"
#> [1] "3.868e-04"
#> [1] ""
#> [1] 325
#> [1] "9.915e-01" "9.834e-01" "9.666e-01"
#> [1] "3.868e-04"
#> [1] ""
#> [1] 326
#> [1] "9.915e-01" "9.834e-01" "9.666e-01"
#> [1] "3.868e-04"
#> [1] ""
#> [1] 327
#> [1] "9.952e-01" "9.896e-01" "9.788e-01"
#> [1] "2.332e-04"
#> [1] ""
#> [1] 328
#> [1] "9.990e-01" "9.986e-01" "9.965e-01"
#> [1] "1.005e-04"
#> [1] ""
#> [1] 329
#> [1] "9.990e-01" "9.986e-01" "9.965e-01"
#> [1] "1.005e-04"
#> [1] ""
#> [1] 330
#> [1] "9.990e-01" "9.986e-01" "9.965e-01"
#> [1] "1.005e-04"
#> [1] ""
#> [1] 331
#> [1] "9.990e-01" "9.986e-01" "9.965e-01"
#> [1] "1.005e-04"
#> [1] ""
#> [1] 332
#> [1] "9.986e-01" "9.979e-01" "9.951e-01"
#> [1] "8.310e-05"
#> [1] ""
#> [1] 333
#> [1] "9.973e-01" "9.947e-01" "9.888e-01"
#> [1] "6.672e-05"
#> [1] ""
#> [1] 334
#> [1] "9.973e-01" "9.947e-01" "9.888e-01"
#> [1] "6.672e-05"
#> [1] ""
#> [1] 335
#> [1] "9.973e-01" "9.947e-01" "9.888e-01"
#> [1] "6.672e-05"
#> [1] ""
#> [1] 336
#> [1] "9.978e-01" "9.959e-01" "9.912e-01"
#> [1] "6.192e-05"
#> [1] ""
#> [1] 337
#> [1] "9.978e-01" "9.959e-01" "9.912e-01"
#> [1] "6.192e-05"
#> [1] ""
#> [1] 338
#> [1] "9.978e-01" "9.959e-01" "9.912e-01"
#> [1] "6.192e-05"
#> [1] ""
#> [1] 339
#> [1] "9.978e-01" "9.959e-01" "9.912e-01"
#> [1] "6.192e-05"
#> [1] ""
#> [1] 340
#> [1] "9.969e-01" "9.938e-01" "9.874e-01"
#> [1] "5.575e-05"
#> [1] ""
#> [1] 341
#> [1] "9.969e-01" "9.938e-01" "9.874e-01"
#> [1] "5.575e-05"
#> [1] ""
#> [1] 342
#> [1] "9.973e-01" "9.948e-01" "9.894e-01"
#> [1] "5.038e-05"
#> [1] ""
#> [1] 343
#> [1] "9.969e-01" "9.940e-01" "9.880e-01"
#> [1] "4.666e-05"
#> [1] ""
#> [1] 344
#> [1] "9.973e-01" "9.947e-01" "9.892e-01"
#> [1] "4.086e-05"
#> [1] ""
#> [1] 345
#> [1] "9.975e-01" "9.950e-01" "9.901e-01"
#> [1] "3.128e-05"
#> [1] ""
#> [1] 346
#> [1] "9.975e-01" "9.950e-01" "9.901e-01"
#> [1] "3.128e-05"
#> [1] ""
#> [1] 347
#> [1] "9.977e-01" "9.956e-01" "9.914e-01"
#> [1] "2.897e-05"
#> [1] ""
#> [1] 348
#> [1] "9.991e-01" "9.986e-01" "9.968e-01"
#> [1] "2.761e-05"
#> [1] ""
#> [1] 349
#> [1] "9.991e-01" "9.986e-01" "9.968e-01"
#> [1] "2.761e-05"
#> [1] ""
#> [1] 350
#> [1] "9.984e-01" "9.971e-01" "9.939e-01"
#> [1] "2.716e-05"
#> [1] ""
#> [1] 351
#> [1] "9.986e-01" "9.970e-01" "9.938e-01"
#> [1] "2.294e-05"
#> [1] ""
#> [1] 352
#> [1] "9.985e-01" "9.970e-01" "9.940e-01"
#> [1] "1.121e-05"
#> [1] ""
#> [1] 353
#> [1] "9.986e-01" "9.972e-01" "9.944e-01"
#> [1] "1.045e-05"
#> [1] ""
#> [1] 354
#> [1] "9.994e-01" "9.990e-01" "9.979e-01"
#> [1] "7.126e-06"