Communicator class

a class for communicating via a temporary file

Public fields

file

is a file which contains the current status

result

is a file in which data can be written or read information.

Methods

Public methods

Communicator$new()

Communicator$getStatus()

Communicator$setStatus()

Communicator$setData()

Communicator$getData()

Communicator$interrupt()

Communicator$ready()

Communicator$running()

Communicator$isInterrupted()

Communicator$destroy()

Communicator$clone()

Method new()

create a new Communicator Object

Usage

Communicator$new()

Method getStatus()

get the current status

Usage

Communicator$getStatus()

Method setStatus()

write a status to the status file

Usage

Communicator$setStatus(msg)

Arguments

msg

is the message which should be set in the file

Method setData()

write data to the result file

Usage

Communicator$setData(data)

Arguments

data

is a string which should be written to the result file

Method getData()

get the current data from the result file

Usage

Communicator$getData()

Method interrupt()

set the status to "interrupt"

Usage

Communicator$interrupt()

Method ready()

set the status to "ready"

Usage

Communicator$ready()

Method running()

write a status to the status file.

Usage

Communicator$running(percComplete)

Arguments

percComplete

is the message which should be set in the file. If percComplete is not passed than the message is set to "Running..."

Method isInterrupted()

Checks if the current status is "interrupt"

Usage

Communicator$isInterrupted()

Method destroy()

removes the temporary files. This method has to be called at the end of the lifetime of the object!

Usage

Communicator$destroy()

Method clone()

The objects of this class are cloneable with this method.

Usage

Communicator$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.