Particle swarm optimization

Interface to the particle swarm optimization (pso) algorithm.

pso(
  env,
  lb,
  ub,
  loss,
  ngen,
  npop,
  error_threshold,
  global = FALSE,
  saveSwarm = FALSE,
  runAsShiny = FALSE
)

Arguments

env

is something that is passed to the loss function in addition to the parameters which get optimized

lb

is a numeric vector defining the lower boundaries of the parameter

ub

is a numeric vector defining the upper boundaries of the parameter

loss

is the loss function for which the optimal parameter set should be found

ngen

is the number of generations the pso should run

npop

is the number of particles which should be used during optimization

error_threshold

is a number defining a sufficient small error at which the optimization is stopped

global

is a logical parameter. If set to TRUE a global star topology is used. Thus, each particle compares itself with the global best particle of the entire swarm. In contrast, if global is set to FALSE a random arbitrary neighborhood is used instead. Thus, each particle has a neighberhood which contains K neighbours where K is between 0 and 3. From the swarm K neighbours are drawn randomly. From the neighberhood the best particle is used for comparison. The neighberhood is calculated for each generation.

saveSwarm

is a logical value defining whether the entire optimization should be saved.

runAsShiny

is an internal parameter which is used when running the shiny app interface.

Examples

rosenbrock <- function(parameter, env, Ignore) {
 value <- 0
 for (i in 1:(length(parameter) - 1)) {
   value <- value + 
     100*(parameter[i + 1] - parameter[i]^2)^2 +
     (1 - parameter[i])^2
 }
return(value)
}

set.seed(1234)
res <- tsf::pso(new.env(), rep(-10, 3), rep(10, 3), rosenbrock, 1000, 40,
                 0.00001, TRUE, FALSE)
#> [1] 2
#> [1] -1.1652582 -0.5884990  0.1541764
#> [1] 389.7224
#> [1] 3
#> [1] -1.880774  1.883203  4.069510
#> [1] 310.0452
#> [1] 4
#> [1] -0.8802364  1.1102363  2.4771296
#> [1] 169.6773
#> [1] 5
#> [1] -1.441102  1.597157  2.412114
#> [1] 31.24546
#> [1] 6
#> [1] -1.441102  1.597157  2.412114
#> [1] 31.24546
#> [1] 7
#> [1] -1.441102  1.597157  2.412114
#> [1] 31.24546
#> [1] 8
#> [1] -1.441102  1.597157  2.412114
#> [1] 31.24546
#> [1] 9
#> [1] -1.441102  1.597157  2.412114
#> [1] 31.24546
#> [1] 10
#> [1] -1.441102  1.597157  2.412114
#> [1] 31.24546
#> [1] 11
#> [1] -1.441102  1.597157  2.412114
#> [1] 31.24546
#> [1] 12
#> [1] -1.441102  1.597157  2.412114
#> [1] 31.24546
#> [1] 13
#> [1] -1.441102  1.597157  2.412114
#> [1] 31.24546
#> [1] 14
#> [1] -1.441102  1.597157  2.412114
#> [1] 31.24546
#> [1] 15
#> [1] -1.441102  1.597157  2.412114
#> [1] 31.24546
#> [1] 16
#> [1] -1.169643  1.733647  2.720230
#> [1] 26.75048
#> [1] 17
#> [1] -1.169643  1.733647  2.720230
#> [1] 26.75048
#> [1] 18
#> [1] -1.169643  1.733647  2.720230
#> [1] 26.75048
#> [1] 19
#> [1] -1.169643  1.733647  2.720230
#> [1] 26.75048
#> [1] 20
#> [1] -1.169643  1.733647  2.720230
#> [1] 26.75048
#> [1] 21
#> [1] -1.169643  1.733647  2.720230
#> [1] 26.75048
#> [1] 22
#> [1] -1.169643  1.733647  2.720230
#> [1] 26.75048
#> [1] 23
#> [1] -1.169643  1.733647  2.720230
#> [1] 26.75048
#> [1] 24
#> [1] -1.169643  1.733647  2.720230
#> [1] 26.75048
#> [1] 25
#> [1] -1.169643  1.733647  2.720230
#> [1] 26.75048
#> [1] 26
#> [1] -1.169643  1.733647  2.720230
#> [1] 26.75048
#> [1] 27
#> [1] -1.169643  1.733647  2.720230
#> [1] 26.75048
#> [1] 28
#> [1] -1.169643  1.733647  2.720230
#> [1] 26.75048
#> [1] 29
#> [1] -1.169643  1.733647  2.720230
#> [1] 26.75048
#> [1] 30
#> [1] -1.169643  1.733647  2.720230
#> [1] 26.75048
#> [1] 31
#> [1] -1.169643  1.733647  2.720230
#> [1] 26.75048
#> [1] 32
#> [1] -1.169643  1.733647  2.720230
#> [1] 26.75048
#> [1] 33
#> [1] -1.169643  1.733647  2.720230
#> [1] 26.75048
#> [1] 34
#> [1] -1.169643  1.733647  2.720230
#> [1] 26.75048
#> [1] 35
#> [1] -1.169643  1.733647  2.720230
#> [1] 26.75048
#> [1] 36
#> [1] -1.169643  1.733647  2.720230
#> [1] 26.75048
#> [1] 37
#> [1] -1.169643  1.733647  2.720230
#> [1] 26.75048
#> [1] 38
#> [1] -1.169643  1.733647  2.720230
#> [1] 26.75048
#> [1] 39
#> [1] -1.169643  1.733647  2.720230
#> [1] 26.75048
#> [1] 40
#> [1] -1.169643  1.733647  2.720230
#> [1] 26.75048
#> [1] 41
#> [1] -1.169643  1.733647  2.720230
#> [1] 26.75048
#> [1] 42
#> [1] -1.339895  1.982703  4.020139
#> [1] 10.74472
#> [1] 43
#> [1] -0.02782453 -0.15137263  0.12792397
#> [1] 5.799664
#> [1] 44
#> [1] -0.02782453 -0.15137263  0.12792397
#> [1] 5.799664
#> [1] 45
#> [1] -0.02782453 -0.15137263  0.12792397
#> [1] 5.799664
#> [1] 46
#> [1] -0.02782453 -0.15137263  0.12792397
#> [1] 5.799664
#> [1] 47
#> [1] -0.02782453 -0.15137263  0.12792397
#> [1] 5.799664
#> [1] 48
#> [1] -0.02782453 -0.15137263  0.12792397
#> [1] 5.799664
#> [1] 49
#> [1] -0.02782453 -0.15137263  0.12792397
#> [1] 5.799664
#> [1] 50
#> [1] -0.02782453 -0.15137263  0.12792397
#> [1] 5.799664
#> [1] 51
#> [1] -0.02782453 -0.15137263  0.12792397
#> [1] 5.799664
#> [1] 52
#> [1] -0.02782453 -0.15137263  0.12792397
#> [1] 5.799664
#> [1] 53
#> [1] -0.02782453 -0.15137263  0.12792397
#> [1] 5.799664
#> [1] 54
#> [1] -0.02782453 -0.15137263  0.12792397
#> [1] 5.799664
#> [1] 55
#> [1] -0.02782453 -0.15137263  0.12792397
#> [1] 5.799664
#> [1] 56
#> [1] -0.02782453 -0.15137263  0.12792397
#> [1] 5.799664
#> [1] 57
#> [1] -0.02782453 -0.15137263  0.12792397
#> [1] 5.799664
#> [1] 58
#> [1] -0.02782453 -0.15137263  0.12792397
#> [1] 5.799664
#> [1] 59
#> [1] -0.02782453 -0.15137263  0.12792397
#> [1] 5.799664
#> [1] 60
#> [1] -0.02782453 -0.15137263  0.12792397
#> [1] 5.799664
#> [1] 61
#> [1] -0.02782453 -0.15137263  0.12792397
#> [1] 5.799664
#> [1] 62
#> [1] -0.02782453 -0.15137263  0.12792397
#> [1] 5.799664
#> [1] 63
#> [1] -0.02782453 -0.15137263  0.12792397
#> [1] 5.799664
#> [1] 64
#> [1] -0.02782453 -0.15137263  0.12792397
#> [1] 5.799664
#> [1] 65
#> [1] -0.02782453 -0.15137263  0.12792397
#> [1] 5.799664
#> [1] 66
#> [1] -0.02782453 -0.15137263  0.12792397
#> [1] 5.799664
#> [1] 67
#> [1] -0.02782453 -0.15137263  0.12792397
#> [1] 5.799664
#> [1] 68
#> [1] -0.02782453 -0.15137263  0.12792397
#> [1] 5.799664
#> [1] 69
#> [1] -0.02782453 -0.15137263  0.12792397
#> [1] 5.799664
#> [1] 70
#> [1] -0.02782453 -0.15137263  0.12792397
#> [1] 5.799664
#> [1] 71
#> [1] -0.02782453 -0.15137263  0.12792397
#> [1] 5.799664
#> [1] 72
#> [1] -0.67590795  0.33227057  0.03248905
#> [1] 5.413642
#> [1] 73
#> [1] -0.67590795  0.33227057  0.03248905
#> [1] 5.413642
#> [1] 74
#> [1] -0.8013542  0.7268291  0.4491597
#> [1] 4.66225
#> [1] 75
#> [1]  0.04405099 -0.10005823  0.12640962
#> [1] 4.519189
#> [1] 76
#> [1]  0.04405099 -0.10005823  0.12640962
#> [1] 4.519189
#> [1] 77
#> [1] -0.5015956  0.2624842  0.1854016
#> [1] 4.167879
#> [1] 78
#> [1] -0.8204712  0.6180357  0.3868113
#> [1] 3.766369
#> [1] 79
#> [1] -0.8204712  0.6180357  0.3868113
#> [1] 3.766369
#> [1] 80
#> [1] -0.7330811  0.5256944  0.3054867
#> [1] 3.327124
#> [1] 81
#> [1] -0.7330811  0.5256944  0.3054867
#> [1] 3.327124
#> [1] 82
#> [1] -0.7330811  0.5256944  0.3054867
#> [1] 3.327124
#> [1] 83
#> [1] -0.7330811  0.5256944  0.3054867
#> [1] 3.327124
#> [1] 84
#> [1] 0.57163431 0.23346080 0.08582802
#> [1] 1.739781
#> [1] 85
#> [1] 0.57163431 0.23346080 0.08582802
#> [1] 1.739781
#> [1] 86
#> [1] 0.6936675 0.4349248 0.1325737
#> [1] 0.9472502
#> [1] 87
#> [1] 0.6936675 0.4349248 0.1325737
#> [1] 0.9472502
#> [1] 88
#> [1] 0.6550940 0.4296443 0.1211927
#> [1] 0.8462665
#> [1] 89
#> [1] 0.6550940 0.4296443 0.1211927
#> [1] 0.8462665
#> [1] 90
#> [1] 0.6550940 0.4296443 0.1211927
#> [1] 0.8462665
#> [1] 91
#> [1] 0.6550940 0.4296443 0.1211927
#> [1] 0.8462665
#> [1] 92
#> [1] 0.6550940 0.4296443 0.1211927
#> [1] 0.8462665
#> [1] 93
#> [1] 0.6550940 0.4296443 0.1211927
#> [1] 0.8462665
#> [1] 94
#> [1] 0.6550940 0.4296443 0.1211927
#> [1] 0.8462665
#> [1] 95
#> [1] 0.6550940 0.4296443 0.1211927
#> [1] 0.8462665
#> [1] 96
#> [1] 0.6550940 0.4296443 0.1211927
#> [1] 0.8462665
#> [1] 97
#> [1] 0.6550940 0.4296443 0.1211927
#> [1] 0.8462665
#> [1] 98
#> [1] 0.6550940 0.4296443 0.1211927
#> [1] 0.8462665
#> [1] 99
#> [1] 0.6550940 0.4296443 0.1211927
#> [1] 0.8462665
#> [1] 100
#> [1] 0.6550940 0.4296443 0.1211927
#> [1] 0.8462665
#> [1] 101
#> [1] 0.6550940 0.4296443 0.1211927
#> [1] 0.8462665
#> [1] 102
#> [1] 0.6550940 0.4296443 0.1211927
#> [1] 0.8462665
#> [1] 103
#> [1] 0.6550940 0.4296443 0.1211927
#> [1] 0.8462665
#> [1] 104
#> [1] 0.6550940 0.4296443 0.1211927
#> [1] 0.8462665
#> [1] 105
#> [1] 0.6550940 0.4296443 0.1211927
#> [1] 0.8462665
#> [1] 106
#> [1] 0.6550940 0.4296443 0.1211927
#> [1] 0.8462665
#> [1] 107
#> [1] 0.6550940 0.4296443 0.1211927
#> [1] 0.8462665
#> [1] 108
#> [1] 0.6550940 0.4296443 0.1211927
#> [1] 0.8462665
#> [1] 109
#> [1] 0.6550940 0.4296443 0.1211927
#> [1] 0.8462665
#> [1] 110
#> [1] 0.6550940 0.4296443 0.1211927
#> [1] 0.8462665
#> [1] 111
#> [1] 0.6550940 0.4296443 0.1211927
#> [1] 0.8462665
#> [1] 112
#> [1] 0.6550940 0.4296443 0.1211927
#> [1] 0.8462665
#> [1] 113
#> [1] 0.6550940 0.4296443 0.1211927
#> [1] 0.8462665
#> [1] 114
#> [1] 0.6550940 0.4296443 0.1211927
#> [1] 0.8462665
#> [1] 115
#> [1] 0.6550940 0.4296443 0.1211927
#> [1] 0.8462665
#> [1] 116
#> [1] 0.6550940 0.4296443 0.1211927
#> [1] 0.8462665
#> [1] 117
#> [1] 0.6550940 0.4296443 0.1211927
#> [1] 0.8462665
#> [1] 118
#> [1] 0.6550940 0.4296443 0.1211927
#> [1] 0.8462665
#> [1] 119
#> [1] 0.6550940 0.4296443 0.1211927
#> [1] 0.8462665
#> [1] 120
#> [1] 0.6550940 0.4296443 0.1211927
#> [1] 0.8462665
#> [1] 121
#> [1] 0.7474053 0.5475328 0.2571795
#> [1] 0.4623956
#> [1] 122
#> [1] 0.7474053 0.5475328 0.2571795
#> [1] 0.4623956
#> [1] 123
#> [1] 0.7474053 0.5475328 0.2571795
#> [1] 0.4623956
#> [1] 124
#> [1] 0.7474053 0.5475328 0.2571795
#> [1] 0.4623956
#> [1] 125
#> [1] 0.7474053 0.5475328 0.2571795
#> [1] 0.4623956
#> [1] 126
#> [1] 0.7474053 0.5475328 0.2571795
#> [1] 0.4623956
#> [1] 127
#> [1] 0.7474053 0.5475328 0.2571795
#> [1] 0.4623956
#> [1] 128
#> [1] 0.7474053 0.5475328 0.2571795
#> [1] 0.4623956
#> [1] 129
#> [1] 0.7474053 0.5475328 0.2571795
#> [1] 0.4623956
#> [1] 130
#> [1] 0.7474053 0.5475328 0.2571795
#> [1] 0.4623956
#> [1] 131
#> [1] 0.7474053 0.5475328 0.2571795
#> [1] 0.4623956
#> [1] 132
#> [1] 0.7474053 0.5475328 0.2571795
#> [1] 0.4623956
#> [1] 133
#> [1] 0.7474053 0.5475328 0.2571795
#> [1] 0.4623956
#> [1] 134
#> [1] 0.7474053 0.5475328 0.2571795
#> [1] 0.4623956
#> [1] 135
#> [1] 0.7474053 0.5475328 0.2571795
#> [1] 0.4623956
#> [1] 136
#> [1] 0.7474053 0.5475328 0.2571795
#> [1] 0.4623956
#> [1] 137
#> [1] 0.7474053 0.5475328 0.2571795
#> [1] 0.4623956
#> [1] 138
#> [1] 0.7474053 0.5475328 0.2571795
#> [1] 0.4623956
#> [1] 139
#> [1] 0.7474053 0.5475328 0.2571795
#> [1] 0.4623956
#> [1] 140
#> [1] 0.7474053 0.5475328 0.2571795
#> [1] 0.4623956
#> [1] 141
#> [1] 0.7474053 0.5475328 0.2571795
#> [1] 0.4623956
#> [1] 142
#> [1] 0.7474053 0.5475328 0.2571795
#> [1] 0.4623956
#> [1] 143
#> [1] 0.7474053 0.5475328 0.2571795
#> [1] 0.4623956
#> [1] 144
#> [1] 0.7474053 0.5475328 0.2571795
#> [1] 0.4623956
#> [1] 145
#> [1] 0.7474053 0.5475328 0.2571795
#> [1] 0.4623956
#> [1] 146
#> [1] 0.7474053 0.5475328 0.2571795
#> [1] 0.4623956
#> [1] 147
#> [1] 0.7474053 0.5475328 0.2571795
#> [1] 0.4623956
#> [1] 148
#> [1] 0.7474053 0.5475328 0.2571795
#> [1] 0.4623956
#> [1] 149
#> [1] 0.7474053 0.5475328 0.2571795
#> [1] 0.4623956
#> [1] 150
#> [1] 0.7474053 0.5475328 0.2571795
#> [1] 0.4623956
#> [1] 151
#> [1] 0.7474053 0.5475328 0.2571795
#> [1] 0.4623956
#> [1] 152
#> [1] 0.7474053 0.5475328 0.2571795
#> [1] 0.4623956
#> [1] 153
#> [1] 0.7474053 0.5475328 0.2571795
#> [1] 0.4623956
#> [1] 154
#> [1] 0.7474053 0.5475328 0.2571795
#> [1] 0.4623956
#> [1] 155
#> [1] 0.7474053 0.5475328 0.2571795
#> [1] 0.4623956
#> [1] 156
#> [1] 0.7474053 0.5475328 0.2571795
#> [1] 0.4623956
#> [1] 157
#> [1] 0.7474053 0.5475328 0.2571795
#> [1] 0.4623956
#> [1] 158
#> [1] 0.7474053 0.5475328 0.2571795
#> [1] 0.4623956
#> [1] 159
#> [1] 0.7850834 0.5765447 0.3405623
#> [1] 0.3906531
#> [1] 160
#> [1] 0.7850834 0.5765447 0.3405623
#> [1] 0.3906531
#> [1] 161
#> [1] 0.7850834 0.5765447 0.3405623
#> [1] 0.3906531
#> [1] 162
#> [1] 0.7850834 0.5765447 0.3405623
#> [1] 0.3906531
#> [1] 163
#> [1] 0.7850834 0.5765447 0.3405623
#> [1] 0.3906531
#> [1] 164
#> [1] 0.7850834 0.5765447 0.3405623
#> [1] 0.3906531
#> [1] 165
#> [1] 0.7850834 0.5765447 0.3405623
#> [1] 0.3906531
#> [1] 166
#> [1] 0.7850834 0.5765447 0.3405623
#> [1] 0.3906531
#> [1] 167
#> [1] 0.7850834 0.5765447 0.3405623
#> [1] 0.3906531
#> [1] 168
#> [1] 0.7850834 0.5765447 0.3405623
#> [1] 0.3906531
#> [1] 169
#> [1] 0.7850834 0.5765447 0.3405623
#> [1] 0.3906531
#> [1] 170
#> [1] 0.7850834 0.5765447 0.3405623
#> [1] 0.3906531
#> [1] 171
#> [1] 0.7850834 0.5765447 0.3405623
#> [1] 0.3906531
#> [1] 172
#> [1] 0.7850834 0.5765447 0.3405623
#> [1] 0.3906531
#> [1] 173
#> [1] 0.7850834 0.5765447 0.3405623
#> [1] 0.3906531
#> [1] 174
#> [1] 0.7850834 0.5765447 0.3405623
#> [1] 0.3906531
#> [1] 175
#> [1] 0.7850834 0.5765447 0.3405623
#> [1] 0.3906531
#> [1] 176
#> [1] 0.7850834 0.5765447 0.3405623
#> [1] 0.3906531
#> [1] 177
#> [1] 0.7850834 0.5765447 0.3405623
#> [1] 0.3906531
#> [1] 178
#> [1] 0.7850834 0.5765447 0.3405623
#> [1] 0.3906531
#> [1] 179
#> [1] 0.7850834 0.5765447 0.3405623
#> [1] 0.3906531
#> [1] 180
#> [1] 0.7850834 0.5765447 0.3405623
#> [1] 0.3906531
#> [1] 181
#> [1] 0.7850834 0.5765447 0.3405623
#> [1] 0.3906531
#> [1] 182
#> [1] 0.7850834 0.5765447 0.3405623
#> [1] 0.3906531
#> [1] 183
#> [1] 0.7850834 0.5765447 0.3405623
#> [1] 0.3906531
#> [1] 184
#> [1] 0.7850834 0.5765447 0.3405623
#> [1] 0.3906531
#> [1] 185
#> [1] 0.7850834 0.5765447 0.3405623
#> [1] 0.3906531
#> [1] 186
#> [1] 0.7850834 0.5765447 0.3405623
#> [1] 0.3906531
#> [1] 187
#> [1] 0.7850834 0.5765447 0.3405623
#> [1] 0.3906531
#> [1] 188
#> [1] 0.7850834 0.5765447 0.3405623
#> [1] 0.3906531
#> [1] 189
#> [1] 0.7850834 0.5765447 0.3405623
#> [1] 0.3906531
#> [1] 190
#> [1] 0.7850834 0.5765447 0.3405623
#> [1] 0.3906531
#> [1] 191
#> [1] 0.7850834 0.5765447 0.3405623
#> [1] 0.3906531
#> [1] 192
#> [1] 0.7850834 0.5765447 0.3405623
#> [1] 0.3906531
#> [1] 193
#> [1] 0.7850834 0.5765447 0.3405623
#> [1] 0.3906531
#> [1] 194
#> [1] 0.7850834 0.5765447 0.3405623
#> [1] 0.3906531
#> [1] 195
#> [1] 0.7850834 0.5765447 0.3405623
#> [1] 0.3906531
#> [1] 196
#> [1] 0.7850834 0.5765447 0.3405623
#> [1] 0.3906531
#> [1] 197
#> [1] 0.7850834 0.5765447 0.3405623
#> [1] 0.3906531
#> [1] 198
#> [1] 0.7850834 0.5765447 0.3405623
#> [1] 0.3906531
#> [1] 199
#> [1] 0.7850834 0.5765447 0.3405623
#> [1] 0.3906531
#> [1] 200
#> [1] 0.7850834 0.5765447 0.3405623
#> [1] 0.3906531
#> [1] 201
#> [1] 0.7850834 0.5765447 0.3405623
#> [1] 0.3906531
#> [1] 202
#> [1] 0.7850834 0.5765447 0.3405623
#> [1] 0.3906531
#> [1] 203
#> [1] 0.7850834 0.5765447 0.3405623
#> [1] 0.3906531
#> [1] 204
#> [1] 0.7850834 0.5765447 0.3405623
#> [1] 0.3906531
#> [1] 205
#> [1] 0.7850834 0.5765447 0.3405623
#> [1] 0.3906531
#> [1] 206
#> [1] 0.7850834 0.5765447 0.3405623
#> [1] 0.3906531
#> [1] 207
#> [1] 0.7850834 0.5765447 0.3405623
#> [1] 0.3906531
#> [1] 208
#> [1] 0.7850834 0.5765447 0.3405623
#> [1] 0.3906531
#> [1] 209
#> [1] 0.7850834 0.5765447 0.3405623
#> [1] 0.3906531
#> [1] 210
#> [1] 0.9012746 0.7918846 0.6623723
#> [1] 0.2192674
#> [1] 211
#> [1] 0.9012746 0.7918846 0.6623723
#> [1] 0.2192674
#> [1] 212
#> [1] 0.9012746 0.7918846 0.6623723
#> [1] 0.2192674
#> [1] 213
#> [1] 0.9012746 0.7918846 0.6623723
#> [1] 0.2192674
#> [1] 214
#> [1] 0.9012746 0.7918846 0.6623723
#> [1] 0.2192674
#> [1] 215
#> [1] 0.9012746 0.7918846 0.6623723
#> [1] 0.2192674
#> [1] 216
#> [1] 0.9012746 0.7918846 0.6623723
#> [1] 0.2192674
#> [1] 217
#> [1] 0.9012746 0.7918846 0.6623723
#> [1] 0.2192674
#> [1] 218
#> [1] 0.9012746 0.7918846 0.6623723
#> [1] 0.2192674
#> [1] 219
#> [1] 0.9012746 0.7918846 0.6623723
#> [1] 0.2192674
#> [1] 220
#> [1] 0.9012746 0.7918846 0.6623723
#> [1] 0.2192674
#> [1] 221
#> [1] 0.9012746 0.7918846 0.6623723
#> [1] 0.2192674
#> [1] 222
#> [1] 0.9012746 0.7918846 0.6623723
#> [1] 0.2192674
#> [1] 223
#> [1] 0.9012746 0.7918846 0.6623723
#> [1] 0.2192674
#> [1] 224
#> [1] 0.9012746 0.7918846 0.6623723
#> [1] 0.2192674
#> [1] 225
#> [1] 0.9012746 0.7918846 0.6623723
#> [1] 0.2192674
#> [1] 226
#> [1] 0.9012746 0.7918846 0.6623723
#> [1] 0.2192674
#> [1] 227
#> [1] 0.9012746 0.7918846 0.6623723
#> [1] 0.2192674
#> [1] 228
#> [1] 0.9035544 0.7941831 0.6646207
#> [1] 0.2159482
#> [1] 229
#> [1] 0.9035544 0.7941831 0.6646207
#> [1] 0.2159482
#> [1] 230
#> [1] 0.9035544 0.7941831 0.6646207
#> [1] 0.2159482
#> [1] 231
#> [1] 0.9035544 0.7941831 0.6646207
#> [1] 0.2159482
#> [1] 232
#> [1] 0.9035544 0.7941831 0.6646207
#> [1] 0.2159482
#> [1] 233
#> [1] 0.9035544 0.7941831 0.6646207
#> [1] 0.2159482
#> [1] 234
#> [1] 0.9035544 0.7941831 0.6646207
#> [1] 0.2159482
#> [1] 235
#> [1] 0.9035544 0.7941831 0.6646207
#> [1] 0.2159482
#> [1] 236
#> [1] 0.9035544 0.7941831 0.6646207
#> [1] 0.2159482
#> [1] 237
#> [1] 0.9035544 0.7941831 0.6646207
#> [1] 0.2159482
#> [1] 238
#> [1] 0.9035544 0.7941831 0.6646207
#> [1] 0.2159482
#> [1] 239
#> [1] 0.9035544 0.7941831 0.6646207
#> [1] 0.2159482
#> [1] 240
#> [1] 0.9035544 0.7941831 0.6646207
#> [1] 0.2159482
#> [1] 241
#> [1] 0.9035544 0.7941831 0.6646207
#> [1] 0.2159482
#> [1] 242
#> [1] 0.9035544 0.7941831 0.6646207
#> [1] 0.2159482
#> [1] 243
#> [1] 0.9035544 0.7941831 0.6646207
#> [1] 0.2159482
#> [1] 244
#> [1] 0.9035544 0.7941831 0.6646207
#> [1] 0.2159482
#> [1] 245
#> [1] 0.9035544 0.7941831 0.6646207
#> [1] 0.2159482
#> [1] 246
#> [1] 0.9035544 0.7941831 0.6646207
#> [1] 0.2159482
#> [1] 247
#> [1] 0.9035544 0.7941831 0.6646207
#> [1] 0.2159482
#> [1] 248
#> [1] 1.0107049 0.9856964 0.9472822
#> [1] 0.1878067
#> [1] 249
#> [1] 1.0107049 0.9856964 0.9472822
#> [1] 0.1878067
#> [1] 250
#> [1] 1.0144178 0.9992523 0.9688480
#> [1] 0.1769146
#> [1] 251
#> [1] 1.0144178 0.9992523 0.9688480
#> [1] 0.1769146
#> [1] 252
#> [1] 1.0144178 0.9992523 0.9688480
#> [1] 0.1769146
#> [1] 253
#> [1] 1.0144178 0.9992523 0.9688480
#> [1] 0.1769146
#> [1] 254
#> [1] 1.0144178 0.9992523 0.9688480
#> [1] 0.1769146
#> [1] 255
#> [1] 1.0144178 0.9992523 0.9688480
#> [1] 0.1769146
#> [1] 256
#> [1] 1.0144178 0.9992523 0.9688480
#> [1] 0.1769146
#> [1] 257
#> [1] 1.0096423 1.0072135 0.9895964
#> [1] 0.07685624
#> [1] 258
#> [1] 1.0096423 1.0072135 0.9895964
#> [1] 0.07685624
#> [1] 259
#> [1] 1.0096423 1.0072135 0.9895964
#> [1] 0.07685624
#> [1] 260
#> [1] 1.0096423 1.0072135 0.9895964
#> [1] 0.07685624
#> [1] 261
#> [1] 1.0096423 1.0072135 0.9895964
#> [1] 0.07685624
#> [1] 262
#> [1] 1.0096423 1.0072135 0.9895964
#> [1] 0.07685624
#> [1] 263
#> [1] 1.0096423 1.0072135 0.9895964
#> [1] 0.07685624
#> [1] 264
#> [1] 1.0096423 1.0072135 0.9895964
#> [1] 0.07685624
#> [1] 265
#> [1] 0.9847857 0.9769981 0.9581252
#> [1] 0.007233601
#> [1] 266
#> [1] 0.9847857 0.9769981 0.9581252
#> [1] 0.007233601
#> [1] 267
#> [1] 0.9872630 0.9796768 0.9605570
#> [1] 0.00312625
#> [1] 268
#> [1] 0.9872630 0.9796768 0.9605570
#> [1] 0.00312625
#> [1] 269
#> [1] 0.9872630 0.9796768 0.9605570
#> [1] 0.00312625
#> [1] 270
#> [1] 0.9872630 0.9796768 0.9605570
#> [1] 0.00312625
#> [1] 271
#> [1] 0.9872630 0.9796768 0.9605570
#> [1] 0.00312625
#> [1] 272
#> [1] 0.9872630 0.9796768 0.9605570
#> [1] 0.00312625
#> [1] 273
#> [1] 0.9872630 0.9796768 0.9605570
#> [1] 0.00312625
#> [1] 274
#> [1] 0.9872630 0.9796768 0.9605570
#> [1] 0.00312625
#> [1] 275
#> [1] 0.9872630 0.9796768 0.9605570
#> [1] 0.00312625
#> [1] 276
#> [1] 0.9872630 0.9796768 0.9605570
#> [1] 0.00312625
#> [1] 277
#> [1] 0.9872630 0.9796768 0.9605570
#> [1] 0.00312625
#> [1] 278
#> [1] 0.9872630 0.9796768 0.9605570
#> [1] 0.00312625
#> [1] 279
#> [1] 0.9872630 0.9796768 0.9605570
#> [1] 0.00312625
#> [1] 280
#> [1] 0.9872630 0.9796768 0.9605570
#> [1] 0.00312625
#> [1] 281
#> [1] 0.9872630 0.9796768 0.9605570
#> [1] 0.00312625
#> [1] 282
#> [1] 0.9872630 0.9796768 0.9605570
#> [1] 0.00312625
#> [1] 283
#> [1] 0.9872630 0.9796768 0.9605570
#> [1] 0.00312625
#> [1] 284
#> [1] 0.9872630 0.9796768 0.9605570
#> [1] 0.00312625
#> [1] 285
#> [1] 0.9872630 0.9796768 0.9605570
#> [1] 0.00312625
#> [1] 286
#> [1] 0.9872630 0.9796768 0.9605570
#> [1] 0.00312625
#> [1] 287
#> [1] 0.9872630 0.9796768 0.9605570
#> [1] 0.00312625
#> [1] 288
#> [1] 0.9872630 0.9796768 0.9605570
#> [1] 0.00312625
#> [1] 289
#> [1] 0.9872630 0.9796768 0.9605570
#> [1] 0.00312625
#> [1] 290
#> [1] 0.9872630 0.9796768 0.9605570
#> [1] 0.00312625
#> [1] 291
#> [1] 0.9872630 0.9796768 0.9605570
#> [1] 0.00312625
#> [1] 292
#> [1] 0.9872630 0.9796768 0.9605570
#> [1] 0.00312625
#> [1] 293
#> [1] 0.9872630 0.9796768 0.9605570
#> [1] 0.00312625
#> [1] 294
#> [1] 0.9872630 0.9796768 0.9605570
#> [1] 0.00312625
#> [1] 295
#> [1] 0.9872630 0.9796768 0.9605570
#> [1] 0.00312625
#> [1] 296
#> [1] 0.9872630 0.9796768 0.9605570
#> [1] 0.00312625
#> [1] 297
#> [1] 0.9872630 0.9796768 0.9605570
#> [1] 0.00312625
#> [1] 298
#> [1] 0.9872630 0.9796768 0.9605570
#> [1] 0.00312625
#> [1] 299
#> [1] 0.9872630 0.9796768 0.9605570
#> [1] 0.00312625
#> [1] 300
#> [1] 0.9911393 0.9848115 0.9664971
#> [1] 0.002038151
#> [1] 301
#> [1] 0.9911393 0.9848115 0.9664971
#> [1] 0.002038151
#> [1] 302
#> [1] 0.9911393 0.9848115 0.9664971
#> [1] 0.002038151
#> [1] 303
#> [1] 0.9911393 0.9848115 0.9664971
#> [1] 0.002038151
#> [1] 304
#> [1] 0.9911393 0.9848115 0.9664971
#> [1] 0.002038151
#> [1] 305
#> [1] 0.9911393 0.9848115 0.9664971
#> [1] 0.002038151
#> [1] 306
#> [1] 0.9911393 0.9848115 0.9664971
#> [1] 0.002038151
#> [1] 307
#> [1] 0.9911393 0.9848115 0.9664971
#> [1] 0.002038151
#> [1] 308
#> [1] 0.9911393 0.9848115 0.9664971
#> [1] 0.002038151
#> [1] 309
#> [1] 0.9911393 0.9848115 0.9664971
#> [1] 0.002038151
#> [1] 310
#> [1] 0.9910431 0.9847025 0.9663916
#> [1] 0.00201205
#> [1] 311
#> [1] 0.9910501 0.9814356 0.9623813
#> [1] 0.0005498577
#> [1] 312
#> [1] 0.9904210 0.9809737 0.9627157
#> [1] 0.0004704202
#> [1] 313
#> [1] 0.9904210 0.9809737 0.9627157
#> [1] 0.0004704202
#> [1] 314
#> [1] 0.9904210 0.9809737 0.9627157
#> [1] 0.0004704202
#> [1] 315
#> [1] 0.9914760 0.9834184 0.9666250
#> [1] 0.0003868022
#> [1] 316
#> [1] 0.9914760 0.9834184 0.9666250
#> [1] 0.0003868022
#> [1] 317
#> [1] 0.9914760 0.9834184 0.9666250
#> [1] 0.0003868022
#> [1] 318
#> [1] 0.9914760 0.9834184 0.9666250
#> [1] 0.0003868022
#> [1] 319
#> [1] 0.9914760 0.9834184 0.9666250
#> [1] 0.0003868022
#> [1] 320
#> [1] 0.9914760 0.9834184 0.9666250
#> [1] 0.0003868022
#> [1] 321
#> [1] 0.9914760 0.9834184 0.9666250
#> [1] 0.0003868022
#> [1] 322
#> [1] 0.9914760 0.9834184 0.9666250
#> [1] 0.0003868022
#> [1] 323
#> [1] 0.9914760 0.9834184 0.9666250
#> [1] 0.0003868022
#> [1] 324
#> [1] 0.9914760 0.9834184 0.9666250
#> [1] 0.0003868022
#> [1] 325
#> [1] 0.9914760 0.9834184 0.9666250
#> [1] 0.0003868022
#> [1] 326
#> [1] 0.9914760 0.9834184 0.9666250
#> [1] 0.0003868022
#> [1] 327
#> [1] 0.9952264 0.9895581 0.9788105
#> [1] 0.0002332002
#> [1] 328
#> [1] 0.9989578 0.9986110 0.9965218
#> [1] 0.0001005267
#> [1] 329
#> [1] 0.9989578 0.9986110 0.9965218
#> [1] 0.0001005267
#> [1] 330
#> [1] 0.9989578 0.9986110 0.9965218
#> [1] 0.0001005267
#> [1] 331
#> [1] 0.9989578 0.9986110 0.9965218
#> [1] 0.0001005267
#> [1] 332
#> [1] 0.9986478 0.9978588 0.9950502
#> [1] 8.310219e-05
#> [1] 333
#> [1] 0.9973253 0.9946505 0.9887733
#> [1] 6.672204e-05
#> [1] 334
#> [1] 0.9973253 0.9946505 0.9887733
#> [1] 6.672204e-05
#> [1] 335
#> [1] 0.9973253 0.9946505 0.9887733
#> [1] 6.672204e-05
#> [1] 336
#> [1] 0.9978341 0.9958849 0.9911883
#> [1] 6.192154e-05
#> [1] 337
#> [1] 0.9978341 0.9958849 0.9911883
#> [1] 6.192154e-05
#> [1] 338
#> [1] 0.9978341 0.9958849 0.9911883
#> [1] 6.192154e-05
#> [1] 339
#> [1] 0.9978341 0.9958849 0.9911883
#> [1] 6.192154e-05
#> [1] 340
#> [1] 0.9968895 0.9938369 0.9874314
#> [1] 5.574955e-05
#> [1] 341
#> [1] 0.9968895 0.9938369 0.9874314
#> [1] 5.574955e-05
#> [1] 342
#> [1] 0.9972703 0.9948253 0.9893866
#> [1] 5.037773e-05
#> [1] 343
#> [1] 0.9969361 0.9939855 0.9879930
#> [1] 4.666267e-05
#> [1] 344
#> [1] 0.9972842 0.9946733 0.9891709
#> [1] 4.08623e-05
#> [1] 345
#> [1] 0.9975342 0.9950161 0.9900704
#> [1] 3.127819e-05
#> [1] 346
#> [1] 0.9975342 0.9950161 0.9900704
#> [1] 3.127819e-05
#> [1] 347
#> [1] 0.9977352 0.9955848 0.9913669
#> [1] 2.897366e-05
#> [1] 348
#> [1] 0.9990996 0.9985747 0.9968239
#> [1] 2.760749e-05
#> [1] 349
#> [1] 0.9990996 0.9985747 0.9968239
#> [1] 2.760749e-05
#> [1] 350
#> [1] 0.9984184 0.9971033 0.9939102
#> [1] 2.715517e-05
#> [1] 351
#> [1] 0.9986200 0.9970217 0.9937817
#> [1] 2.294334e-05
#> [1] 352
#> [1] 0.9985078 0.9970060 0.9940122
#> [1] 1.121267e-05
#> [1] 353
#> [1] 0.9986043 0.9971707 0.9944076
#> [1] 1.04494e-05
#> [1] 354
#> [1] 0.9994322 0.9990335 0.9978940
#> [1] 7.126325e-06